+++ packages/api/src/http/api-routes.ts
@@ -69,7 +69,12 @@ import {
   getMapFilters,
   updateMapFilter,
 } from './controllers/map-filters-controllers';
-import { transferResources } from './controllers/marketplace-controllers';
+import {
+  createTradeRoute,
+  deleteTradeRoute,
+  transferResources,
+  updateTradeRoute,
+} from './controllers/marketplace-controllers';
 import { getOasesWithAnimals } from './controllers/oasis-animal-finder-controllers';
 import { getTilesWithBonuses } from './controllers/oasis-bonus-finder-controllers';
 import { abandonOasis, occupyOasis } from './controllers/oasis-controllers';
@@ -250,6 +255,9 @@ export const apiRoutes = [
   createRoute(rearrangeBuildingFields),
   createRoute(getResearchedUnits),
   createRoute(transferResources),
+  createRoute(createTradeRoute),
+  createRoute(updateTradeRoute),
+  createRoute(deleteTradeRoute),
 
   // Map Filters
   createRoute(getMapFilters),